# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Sandbox CESFAM",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Login",
        "description": "Login sandbox: acepta cualquier credencial, devuelve un token stub.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/batches/{batch_id}/write-off": {
      "post": {
        "tags": [
          "Partidas"
        ],
        "summary": "Write Off",
        "description": "Da de baja unidades de una partida. Descuenta availableQuantity siempre;\ndescuenta physicalQuantity solo cuando discard=true (requerimiento de auditoría).\nPersiste un registro WriteOff con motivo, cantidad, partida, medicamento y staff.",
        "operationId": "write_off_api_v1_batches__batch_id__write_off_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchWriteOff"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Batch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/write-offs": {
      "get": {
        "tags": [
          "Bajas"
        ],
        "summary": "List Write Offs",
        "description": "Lista bajas registradas. Soporta filtros por medicamento, partida, estado y rango de fechas.",
        "operationId": "list_write_offs_api_v1_write_offs_get",
        "parameters": [
          {
            "name": "medicationId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Medicationid"
            }
          },
          {
            "name": "batchId",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Batchid"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datefrom"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dateto"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/write-offs/{write_off_id}": {
      "get": {
        "tags": [
          "Bajas"
        ],
        "summary": "Get Write Off",
        "operationId": "get_write_off_api_v1_write_offs__write_off_id__get",
        "parameters": [
          {
            "name": "write_off_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Write Off Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteOff"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Batch": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "medicationId": {
            "type": "string",
            "title": "Medicationid"
          },
          "batchNumber": {
            "type": "string",
            "title": "Batchnumber"
          },
          "arrivalDate": {
            "type": "string",
            "format": "date",
            "title": "Arrivaldate"
          },
          "expirationDate": {
            "type": "string",
            "format": "date",
            "title": "Expirationdate"
          },
          "initialQuantity": {
            "type": "integer",
            "title": "Initialquantity"
          },
          "availableQuantity": {
            "type": "integer",
            "title": "Availablequantity"
          }
        },
        "type": "object",
        "required": [
          "id",
          "medicationId",
          "batchNumber",
          "arrivalDate",
          "expirationDate",
          "initialQuantity",
          "availableQuantity"
        ],
        "title": "Batch"
      },
...
      "BatchWriteOff": {
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/WriteOffReason"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          },
          "discard": {
            "type": "boolean",
            "title": "Discard",
            "default": false
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "quantity"
        ],
        "title": "BatchWriteOff"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "title": "LoginResponse"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "rut": {
            "type": "string",
            "title": "Rut"
          },
          "fullName": {
            "type": "string",
            "title": "Fullname"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "rut",
          "fullName",
          "role"
        ],
        "title": "User"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "doctor",
          "pharmacy_staff"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WriteOff": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batchId": {
            "type": "string",
            "title": "Batchid"
          },
          "medicationId": {
            "type": "string",
            "title": "Medicationid"
          },
          "staffId": {
            "type": "string",
            "title": "Staffid"
          },
          "reason": {
            "$ref": "#/components/schemas/WriteOffReason"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "status": {
            "$ref": "#/components/schemas/WriteOffStatus"
          },
          "expiredAt": {
            "type": "string",
            "format": "date",
            "title": "Expiredat"
          },
          "discardDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discarddate"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "batchId",
          "medicationId",
          "staffId",
          "reason",
          "quantity",
          "status",
          "expiredAt"
        ],
        "title": "WriteOff"
      },
      "WriteOffReason": {
        "type": "string",
        "enum": [
          "EXPIRATION",
          "DAMAGED",
          "BROKEN_PACKAGE",
          "OTHER"
        ],
        "title": "WriteOffReason"
      },
      "WriteOffStatus": {
        "type": "string",
        "enum": [
          "DEDUCTED_FROM_AVAILABLE",
          "DISCARDED"
        ],
        "title": "WriteOffStatus"
      }
    }
  }
}